+++ extensions/brave/src/search-bookmarks.tsx
@@ -1,10 +1,13 @@
+import type { ReactElement } from "react";
+import { useEffect } from "react";
+
 import { List } from "@raycast/api";
-import { ReactElement, useEffect } from "react";
-import { BraveListItems } from "./components";
-import { useBookmarkSearch } from "./hooks/useBookmarkSearch";
 import { useCachedState } from "@raycast/utils";
-import { BRAVE_PROFILE_KEY, DEFAULT_BRAVE_PROFILE_ID, NoBookmarksText } from "./constants";
+
+import { BraveListItems } from "./components";
 import BraveProfileDropDown from "./components/BraveProfileDropdown";
+import { BRAVE_PROFILE_KEY, DEFAULT_BRAVE_PROFILE_ID, NoBookmarksText } from "./constants";
+import { useBookmarkSearch } from "./hooks/useBookmarkSearch";
 
 export default function Command(): ReactElement {
   const [profile] = useCachedState(BRAVE_PROFILE_KEY, DEFAULT_BRAVE_PROFILE_ID);
@@ -17,9 +20,7 @@ export default function Command(): ReactElement {
   return (
     <List filtering={true} isLoading={isLoading} searchBarAccessory={<BraveProfileDropDown />}>
       <List.EmptyView title={NoBookmarksText} icon={{ source: "empty-view.png" }} />
-      {data?.map((e) => (
-        <BraveListItems.TabHistory entry={e} key={e.id} profile={profile} />
-      ))}
+      {data?.map((e) => <BraveListItems.TabHistory entry={e} key={e.id} profile={profile} />)}
     </List>
   );
 }