+++ extensions/brave/src/search-history.tsx
@@ -1,11 +1,14 @@
+import type { ReactElement } from "react";
+import { useEffect, useState } from "react";
+
 import { List } from "@raycast/api";
-import { useState, ReactElement, useEffect } from "react";
-import { useHistorySearch } from "./hooks/useHistorySearch";
-import { BraveProfile, GroupedEntries, HistoryEntry } from "./interfaces";
-import { BraveListItems } from "./components";
 import { useCachedState } from "@raycast/utils";
-import { BRAVE_PROFILES_KEY, BRAVE_PROFILE_KEY, DEFAULT_BRAVE_PROFILE_ID } from "./constants";
+
+import { BraveListItems } from "./components";
 import BraveProfileDropDown from "./components/BraveProfileDropdown";
+import { BRAVE_PROFILES_KEY, BRAVE_PROFILE_KEY, DEFAULT_BRAVE_PROFILE_ID } from "./constants";
+import { useHistorySearch } from "./hooks/useHistorySearch";
+import type { BraveProfile, GroupedEntries, HistoryEntry } from "./interfaces";
 
 const groupEntries = (allEntries?: HistoryEntry[]): GroupedEntries =>
   allEntries
@@ -59,9 +62,7 @@ export default function Command(): ReactElement {
     >
       {groups?.map((group) => (
         <List.Section title={group} key={group}>
-          {groupedEntries?.get(group)?.map((e) => (
-            <BraveListItems.TabHistory entry={e} key={e.id} profile={profile} />
-          ))}
+          {groupedEntries?.get(group)?.map((e) => <BraveListItems.TabHistory entry={e} key={e.id} profile={profile} />)}
         </List.Section>
       ))}
     </List>