+++ extensions/brave/src/search-tab.tsx
@@ -1,7 +1,8 @@
-import { getPreferenceValues, List } from "@raycast/api";
+import { List, getPreferenceValues } from "@raycast/api";
+
 import { BraveListItems } from "./components";
 import { useTabSearch } from "./hooks/useTabSearch";
-import { Preferences } from "./interfaces";
+import type { Preferences } from "./interfaces";
 
 export default function Command() {
   const { useOriginalFavicon } = getPreferenceValues<Preferences>();
@@ -13,9 +14,7 @@ export default function Command() {
 
   return (
     <List isLoading={isLoading}>
-      {data?.map((tab) => (
-        <BraveListItems.TabList key={tab.key()} tab={tab} useOriginalFavicon={useOriginalFavicon} />
-      ))}
+      {data?.map((tab) => <BraveListItems.TabList key={tab.key()} tab={tab} useOriginalFavicon={useOriginalFavicon} />)}
     </List>
   );
 }